+++ src/encoder/avi_encoder.ml
@@ -78,8 +78,18 @@ let encoder avi =
       Strings.dda header ans )
     else ans
   in
+  let hls =
+    {
+      Encoder.init_encode = (fun f o l -> (None, encode f o l));
+      split_encode = (fun f o l -> `Ok (Strings.empty, encode f o l));
+      codec_attrs = (fun () -> None);
+      bitrate = (fun () -> None);
+      video_size = (fun () -> None);
+    }
+  in
   {
     Encoder.insert_metadata = (fun _ -> ());
+    hls;
     encode;
     header = Strings.of_string header;
     stop = (fun () -> Strings.empty);