+++ src/encoder/ogg_encoder.ml
@@ -102,7 +102,7 @@ let encoder { Ogg_format.audio; video } =
     let skeleton = List.length tracks > 1 in
     let ogg_enc = Ogg_muxer.create ~skeleton name in
     let rec enc =
-      { Encoder.insert_metadata; encode; header = Strings.empty; stop }
+      { Encoder.insert_metadata; hls; encode; header = Strings.empty; stop }
     and streams_start () =
       let f track =
         match track.id with
@@ -123,6 +123,14 @@ let encoder { Ogg_format.audio; video } =
       in
       List.iter f tracks;
       Ogg_muxer.get_data ogg_enc
+    and hls =
+      {
+        Encoder.init_encode = (fun f o l -> (None, encode f o l));
+        split_encode = (fun f o l -> `Ok (Strings.empty, encode f o l));
+        codec_attrs = (fun () -> None);
+        bitrate = (fun () -> None);
+        video_size = (fun () -> None);
+      }
     and ogg_stop () =
       let f track = track.id <- None in
       List.iter f tracks;